+++ lib/plugins/aws/deployList.js
@@ -1,9 +1,9 @@
 'use strict';
 
 const BbPromise = require('bluebird');
-const validate = require('../lib/validate');
-const findAndGroupDeployments = require('../utils/findAndGroupDeployments');
-const setBucketName = require('../lib/setBucketName');
+const validate = require('./lib/validate');
+const findAndGroupDeployments = require('./utils/findAndGroupDeployments');
+const setBucketName = require('./lib/setBucketName');
 
 class AwsDeployList {
   constructor(serverless, options) {
