+++ lib/plugins/index.js
@@ -53,7 +53,7 @@ module.exports = [
   require('./aws/package/compile/events/cloudFront/index.js'),
   require('./aws/package/compile/events/httpApi'),
   require('./aws/deployFunction.js'),
-  require('./aws/deployList/index.js'),
+  require('./aws/deployList.js'),
   require('./aws/invokeLocal/index.js'),
   require('./standalone'),
 ];
